+++ src/nextToken.re2c.cpp
@@ -320,8 +320,10 @@ INIT:
 
     <CMD> [0-9]* "<"         { RET(REDIR_IN); }
     <CMD> [0-9]* ">"         { RET(REDIR_OUT); }
+    <CMD> [0-9]* ">|"        { RET(REDIR_OUT_CLOBBER); }
     <CMD> [0-9]* ">>"        { RET(REDIR_APPEND); }
     <CMD> [0-9]* "&>"        { RET(REDIR_OUT_ERR); }
+    <CMD> [0-9]* "&>|"       { RET(REDIR_OUT_ERR_CLOBBER); }
     <CMD> [0-9]* "&>>"       { RET(REDIR_APPEND_OUT_ERR); }
     <CMD> [0-9]* "<&"        { RET(REDIR_DUP_IN); }
     <CMD> [0-9]* ">&"        { RET(REDIR_DUP_OUT); }