+++ src/vm.h
@@ -36,15 +36,16 @@ namespace ydsh {
 
 #define EACH_RUNTIME_OPTION(OP)                                                                    \
   OP(ASSERT, (1u << 0u), "assert")                                                                 \
-  OP(DOTGLOB, (1u << 1u), "dotglob")                                                               \
-  OP(ERR_RAISE, (1u << 2u), "errraise")                                                            \
-  OP(FASTGLOB, (1u << 3u), "fastglob")                                                             \
-  OP(FAIL_SIGPIPE, (1u << 4u), "failsigpipe")                                                      \
-  OP(HUP_EXIT, (1u << 5u), "huponexit")                                                            \
-  OP(MONITOR, (1u << 6u), "monitor")                                                               \
-  OP(NULLGLOB, (1u << 7u), "nullglob")                                                             \
-  OP(TRACE_EXIT, (1u << 8u), "traceonexit")                                                        \
-  OP(XTRACE, (1u << 9u), "xtrace")
+  OP(CLOBBER, (1u << 1u), "clobber")                                                               \
+  OP(DOTGLOB, (1u << 2u), "dotglob")                                                               \
+  OP(ERR_RAISE, (1u << 3u), "errraise")                                                            \
+  OP(FASTGLOB, (1u << 4u), "fastglob")                                                             \
+  OP(FAIL_SIGPIPE, (1u << 5u), "failsigpipe")                                                      \
+  OP(HUP_EXIT, (1u << 6u), "huponexit")                                                            \
+  OP(MONITOR, (1u << 7u), "monitor")                                                               \
+  OP(NULLGLOB, (1u << 8u), "nullglob")                                                             \
+  OP(TRACE_EXIT, (1u << 9u), "traceonexit")                                                        \
+  OP(XTRACE, (1u << 10u), "xtrace")
 
 // set/unset via 'shctl' command
 enum class RuntimeOption : unsigned short {
@@ -102,7 +103,8 @@ public:
 
   bool isInteractive{false};
 
-  RuntimeOption runtimeOption{RuntimeOption::HUP_EXIT | RuntimeOption::ASSERT};
+  RuntimeOption runtimeOption{RuntimeOption::HUP_EXIT | RuntimeOption::ASSERT |
+                              RuntimeOption::CLOBBER};
 
   DSExecMode execMode{DS_EXEC_MODE_NORMAL};